requirements.txt
asgiref==3.4.1
Django==3.2.16
django-extra-views==0.14.0
django-filter==21.1
django-widget-tweaks==1.4.11
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
